+++ package.json
@@ -18,7 +18,7 @@
   "license": "MIT",
   "scripts": {
     "dev": "vite build --watch",
-    "build": "tsc --emitDeclarationOnly && vite build",
+    "build": "rm -rf types && tsc --emitDeclarationOnly && vite build",
     "prepublishOnly": "npm run build"
   },
   "dependencies": {},
